requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blis==0.2.4
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cycler==0.10.0
cymem==2.0.5
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
entrypoints==0.3
filelock==3.0.12
future==0.18.2
hyperopt==0.2.5
idna==2.10
importlib-metadata==3.3.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
Jinja2==2.11.2
joblib==1.0.0
jsonlines==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.6
numpy==1.19.5
packaging==20.8
pandocfilters==1.4.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.9.0
ptyprocess==0.6.0
pyaml==20.4.0
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scipy==1.5.4
Send2Trash==1.5.0
sentencepiece==0.1.94
six==1.15.0
spacy==2.1.6
srsly==1.0.5
terminado==0.9.2
testpath==0.4.4
thinc==7.0.8
tokenizers==0.8.0rc4
torch==1.6.0
torchtext==0.4.0
tornado==6.0.4
tqdm==4.55.1
traitlets==4.3.3
transformers==3.0.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
